+++ pepe-tg/src/actions/index.ts
@@ -3,5 +3,6 @@ export { fakeCommonsCardAction } from "./fakeCommonsCard";
 export { rarePepesCardAction } from "./rarePepesCard";
 export { startCommand, helpCommand } from "./basicCommands";
 export { fakeRememberCommand } from "./fakeRememberCommand";
+export { vouchCommand } from "./vouchCommand";
 export { costCommand } from "./costCommand";
 export { xcpCommand } from "./xcpCommand";